+++ lib/gridstore/src/blobstore/arenastore/view.rs
